requirements.txt
Flask==2.3.3
flask-session==0.5.0
python-dotenv==1.0.0
spotipy==2.22.1
requests==2.31.0
